Add files via upload
added option for "print_version" argument (of "print_version_info" function) to be represented as a List containing a bool for first entry (normal) and a str for second entry (default=None). If the str is not None (e.g. 'full') then the python version will print in it's full form

also changed format for template to have an easier-to-read workspace (--begin(), end(), """ Top of Stack""", """Bottom of Stack""" ), 

added version info to README.